Password action only for keycloak
