+++ airflow-core/src/airflow/cli/cli_config.py
@@ -21,7 +21,6 @@
 from __future__ import annotations
 
 import argparse
-import getpass
 import json
 import os
 import textwrap
@@ -140,15 +139,6 @@ def string_lower_type(val):
     if not val:
         return
     return val.strip().lower()
-
-
-class Password(argparse.Action):
-    """Custom action to prompt for password input."""
-
-    def __call__(self, parser, namespace, values, option_string=None):
-        if values is None:
-            values = getpass.getpass(prompt="Password: ")
-        setattr(namespace, self.dest, values)
 
 
 # Shared
