+++ providers/keycloak/tests/unit/keycloak/auth_manager/cli/test_definition.py
@@ -16,13 +16,14 @@
 # under the License.
 from __future__ import annotations
 
+import argparse
 import importlib
 from unittest.mock import patch
 
 import pytest
 
 from airflow.cli import cli_parser
-from airflow.providers.keycloak.auth_manager.cli.definition import KEYCLOAK_AUTH_MANAGER_COMMANDS
+from airflow.providers.keycloak.auth_manager.cli.definition import KEYCLOAK_AUTH_MANAGER_COMMANDS, Password
 
 from tests_common.test_utils.config import conf_vars
 
@@ -76,5 +77,29 @@ class TestKeycloakCliDefinition:
             "--password",
         ]
         args = self.arg_parser.parse_args(params)
+        mock_getpass.assert_called_once_with(prompt="Password: ")
+        assert args.password == "stdin_password"
+
+
+class TestPasswordAction:
+    """Tests for the Password argparse action."""
+
+    def test_password_with_explicit_value(self):
+        """Test passing password explicitly via --password value."""
+
+        parser = argparse.ArgumentParser()
+        parser.add_argument("--password", action=Password, nargs="?", dest="password")
+
+        args = parser.parse_args(["--password", "my_password"])
+        assert args.password == "my_password"
+
+    @patch("getpass.getpass", return_value="stdin_password")
+    def test_password_from_stdin(self, mock_getpass):
+        """Test password prompted from stdin when --password has no value."""
+
+        parser = argparse.ArgumentParser()
+        parser.add_argument("--password", action=Password, nargs="?", dest="password")
+
+        args = parser.parse_args(["--password"])
         mock_getpass.assert_called_once_with(prompt="Password: ")
         assert args.password == "stdin_password"
